Move concept functions to concept_manager
