+++ src/omop_etl_wrapper/model/table_manager/__init__.py
@@ -1,2 +1,3 @@
 from .vocabulary_manager import VocabManager
 from .concept_class_manager import ClassManager
+from .concept_manager import ConceptManager
